Moved get phonetic data out of the default function. Moved processBaybayinText into its own file.
